Make values() really sort as documented
The values() method used to sort the returned list by the update time
(newest first), contrary to the documentation.

This was mostly missed by the test suite, except on rare occasions,
because the time stamps don't use sub-second precision. Only when the
threshold between full seconds happened in the middle of the data
insertion statements, the values() test in t/001Basic.t failed.

Fix the method, change the data insertion statement order, and add a
sleep statement in the middle to make sure the bug stays fixed.